More work on estimators.

+++ TBag.BloomFilters/Estimators/HybridEstimatorDataExtensions.cs
@@ -18,7 +18,7 @@
         /// <param name="configuration">Configuration</param>
         /// <param name="destructive">When <c>true</c> the values of <paramref name="estimator"/> will be altered rendering it useless, otherwise <c>false</c></param>
         /// <returns>An estimate of the difference between two sets based upon the estimators.</returns>
-        public static ulong Decode<TEntity, TId, TCount>(this IHybridEstimatorData<int, TCount> estimator,
+        public static long Decode<TEntity, TId, TCount>(this IHybridEstimatorData<int, TCount> estimator,
             IHybridEstimatorData<int, TCount> otherEstimatorData,
             IBloomFilterConfiguration<TEntity, TId, int, int, TCount> configuration,
              bool destructive = false)
@@ -26,13 +26,13 @@
             where TId : struct
         {
             if (estimator == null && otherEstimatorData == null) return 0L;
-            if (estimator == null) return (ulong)otherEstimatorData.CountEstimate;
-            if (otherEstimatorData == null) return (ulong)estimator.CountEstimate;
+            if (estimator == null) return otherEstimatorData.CountEstimate;
+            if (otherEstimatorData == null) return estimator.CountEstimate;
             var decodeFactor = Math.Max(estimator.StrataEstimator?.DecodeCountFactor ?? 1.0D,
                 otherEstimatorData.StrataEstimator?.DecodeCountFactor ?? 1.0D);
             var maxDifference = otherEstimatorData.CountEstimate + estimator.CountEstimate;
             var strataDecode = estimator.StrataEstimator.Decode(otherEstimatorData.StrataEstimator, configuration, maxDifference, destructive);
-            var minwiseDecode = (ulong)(decodeFactor * (estimator.BitMinwiseEstimator.Capacity - 
+            var minwiseDecode = (long)(decodeFactor * (estimator.BitMinwiseEstimator.Capacity - 
                     estimator.BitMinwiseEstimator.Similarity(otherEstimatorData.BitMinwiseEstimator) * 
                         estimator.BitMinwiseEstimator.Capacity));
             return strataDecode + minwiseDecode;
